% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendlyverse.R
\docType{package}
\name{friendlyverse-package}
\alias{friendlyverse}
\alias{friendlyverse-package}
\title{friendlyverse: Distribution of Friendlyverse Packages}
\description{
The 'friendlyverse' is a set of packages that work in
    harmony because they are all about having fun. This package is designed to
    make it easy to install and load multiple 'friendlyverse' packages in a
    single step. Use your power for good, not evil.
}
\author{
\strong{Maintainer}: Jay Qi \email{jayqi.opensource@gmail.com}

Authors:
\itemize{
  \item Stephanie Kirmer \email{stephanie@stephaniekirmer.com}
}

}
\keyword{internal}
