% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendlyverse.R
\name{friendlyverse_packages}
\alias{friendlyverse_packages}
\title{List friendlyverse packages}
\usage{
friendlyverse_packages()
}
\value{
character vector
}
\description{
List all friendlyverse packages.
}
