% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendlyverse.R
\name{friendlyverse_reinstall}
\alias{friendlyverse_reinstall}
\title{Reinstall friendlyverse packages}
\usage{
friendlyverse_reinstall()
}
\value{
NULL invisibly
}
\description{
Reinstall all friendlyverse packages from CRAN.
}
